Conflicts:
	src/main/java/org/cancure/cpa/persistence/entity/InvoicesEntity.java
	src/main/resources/schema-mysql.sql
	src/main/resources/schema.sql
	src/main/webapp/app/common/appConstant.js

+++ src/main/java/org/cancure/cpa/controller/beans/PatientVisitDocumentBean.java
@@ -4,7 +4,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 public class PatientVisitDocumentBean {
 
-	private Long docId;
+	private Integer docId;
 	private String docType;
 	private String docPath;
 	private Integer accountTypeId; 
@@ -18,11 +18,11 @@ public class PatientVisitDocumentBean {
 		this.patientVisitFile = patientVisitFile;
 	}
 
-	public Long getDocId() {
+	public Integer getDocId() {
 		return docId;
 	}
 
-	public void setDocId(Long docId) {
+	public void setDocId(Integer docId) {
 		this.docId = docId;
 	}
